Update tests to reflect refactor
